smartds/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.12.1
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
celery==5.2.7
certifi==2019.11.28
channels==4.0.0
channels-redis==4.1.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0
dronekit==2.9.2
dronekit-sitl==3.3.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.3
geographiclib==2.0
geopy==2.3.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
loguru==0.6.0
logzero==1.7.0
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
MAVProxy==1.8.22
monotonic==1.6
motor==3.1.2
msgpack==1.0.5
multidict==6.0.4
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
openai==0.27.4
pandas==2.0.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
pinecone-client==2.2.1
prompt-toolkit==3.0.38
protobuf==3.6.1
psutil==5.9.5
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymavlink==2.4.8
pymongo==4.3.3
PyNaCl==1.3.0
PyPDF2==2.10.9
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==1.0.0
pytz==2023.3
pyxdg==0.26
PyYAML==6.0
redis==4.5.4
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rplidar==0.9.2
SecretStorage==2.3.1
sentry-sdk==1.19.1
simplejson==3.16.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
systemd-python==234
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.15
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.6
websockets==11.0.3
xkit==0.0.0
yarl==1.8.2
